\name{mexicomapR-package}
\alias{mexicomapR-package}
\alias{mexicomapR}
\docType{package}
\title{
Data and functions to create Mexican choropleth maps
}
\description{
Easily create Mexican admin-1 (state) and admin-2 (municipality) choropleth maps with self-contained shapefiles and functions. Wrappers for sp, ggplot and leaflet packages.
}
\details{
\tabular{ll}{
Package: \tab mexicomapR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-1-1\cr
License: \tab What license is it under?\cr
}
mxmap calls all other functions.
}
\author{
Eduardo Flores
Maintainer: Eduardo Flores <eduardo@enelmargen.org>
}
\keyword{ package }