% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_functions.R
\name{mx_checkdf}
\alias{mx_checkdf}
\title{Checks and cleans format of data.frame to be merged with fortified data}
\usage{
mx_checkdf(df, vcol = "value", nobject = 32)
}
\arguments{
\item{df}{Data.frame with id and value columns}

\item{vcol}{Value column with standard evaluation. Defaults to "value".}
}
\description{
Checks unique id's, format of value and other possible issues.
}
\examples{

}
\author{
Eduardo Flores
}

