% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_functions.R
\name{mxfind_id}
\alias{mxfind_id}
\title{Attempts to fix INEGI number to description of geography}
\usage{
mxfind_id(vector)
}
\arguments{
\item{vector}{Vector of names or descriptions}
}
\description{
Calls \code{data(mxgeography)} and attempts to translate geography names to INEGI codes.
}
\examples{
\dontrun{
token<-"webservice_token"
CrecimientoZonas<-series_crecimiento_regiones(token)
}
}
\author{
Eduardo Flores
}

