% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{mxgeography}
\alias{mxgeography}
\title{Catalog of geographical names and INEGI equivalents}
\format{A data frame with 9 variables}
\usage{
data(mxgeography)
}
\description{
A data catalog that matches some common descriptors and it's official geographical equivalents from INEGI.
}
\details{
\itemize{
  \item mun. Name of Municipality, as defined by INEGI
  \item mun_id. Municipality id in mxmunicipios
  \item city. City as defined arbitrarily
  \item met. Metropolitan zone, defined arbitrarily, can be matched with names in price index of inegiR (inflacion_ciudades)
  \item name_2. Name of Municipality, as defined internally
  \item st. State name as defined by INEGI
  \item st_name. State namem defined arbitrarily to be matched with inegiR
  \item st_id. State id
  \item zn. Zone, as defined by INEGI: \url{http://www.inegi.org.mx/est/contenidos/proyectos/cn/itaee/default.aspx}.
}
}
\keyword{datasets}

