% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{mxmap}
\alias{mxmap}
\title{Creates a map}
\usage{
mxmap(type, df, method = "static", vcol = "value")
}
\arguments{
\item{type}{Type of map to create (data by state or municipality)}

\item{df}{Data frame or vector with information to paint geography. The data must contain an "id" and a "value" column to be matched.}

\item{method}{Method to be used (static for ggplot2 object, leaflet for leaflet object)}
}
\value{
ggplot object
}
\description{
Calls \code{mxmap_state} or \code{mxmap_municipality} to generate a static map.
}
\examples{
# A static state map with random data 
data(states_catalog, package = "mexicomapR")
df <- data.frame("id" = states_catalog$id, "value" = runif(n = 32, min = 1, max = 100))
mxmap(type = "state", df)
}
\author{
Eduardo Flores
}
\seealso{
mxmap_state_p, mxmap_state_l, mxmap_state_g, mxmap_municipality_p, mxmap_municipality_l, mxmap_municipality_g
}

