% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{mxmap_state_l}
\alias{mxmap_state_l}
\title{Creates leaflet state map}
\usage{
mxmap_state_l(df)
}
\arguments{
\item{df}{Data frame or vector with information to paint geography. The data must contain an "id" and a "value" column to be matched.}
}
\value{
htmlwidget object
}
\description{
Generates a leaflet state map
}
\examples{
\dontrun{
# change... 
population_choroplethr <- mxmap("municipality", data(mxpop2010))
}
}
\author{
Eduardo Flores
}
\seealso{
mxmap
}

