% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{mxmap_state_p}
\alias{mxmap_state_p}
\title{Creates state map}
\usage{
mxmap_state_p(df)
}
\arguments{
\item{df}{Data frame or vector with information to paint geography. The data must contain an "id" and a "value" column to be matched.}
}
\value{
ggplot object
}
\description{
Generates a static state map
}
\examples{
\dontrun{
population_choroplethr <- mxmap("state", data(mxpop2010))
}
}
\author{
Eduardo Flores
}
\seealso{
mxmap
}

