% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{mxmap_state_s}
\alias{mxmap_state_s}
\title{Creates static state map}
\usage{
mxmap_state_s(df, vcol = "value")
}
\arguments{
\item{df}{Data frame or vector with information to paint geography. The data must contain an "id" and a "value" column to be matched.}

\item{vcol}{Value column with standard evaluation. Defaults to "value".}
}
\value{
ggplot object
}
\description{
Generates a static (ggplot2 object) state map.
}
\examples{
# A state map with random data 
utils::data(states_catalog, package = "mexicomapR")
df <- data.frame("id" = states_catalog$id, "value" = runif(n = 32, min = 1, max = 100))
mxmap_state_s(df)
}
\author{
Eduardo Flores
}
\seealso{
mxmap
}

