% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{mxmunicipality}
\alias{mxmunicipality}
\title{Mexican municipalities}
\format{A data frame with 9 variables}
\usage{
data(mxmunicipality)
}
\description{
A fortified data.frame with Mexican municipality borders, to be passed to ggplot.
}
\details{
\itemize{
  \item long. Longitud 
  \item lat. Latitud
  \item group. Group to identify same coordinates
  \item id. Unique municipality id to match data.frame with information (number is a concatenation of INEGI entidad id and municipality id)
}
}
\keyword{datasets}

