% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{mxmunicipality_catalog}
\alias{mxmunicipality_catalog}
\title{Creates municipality catalog}
\usage{
mxmunicipality_catalog()
}
\value{
data.frame
}
\description{
Uses \code{data(mxgeography)} to generate a data.frame with two columns: id as factor and name of municipality, on which to join data. This assures the user will use the correct id's to pass to \code{mxmap()}.
}
\examples{
mxmunicipality_catalog()
}
\author{
Eduardo Flores
}
\seealso{
mxgeography
}

