% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{mxstates}
\alias{mxstates}
\title{Mexican states}
\format{A data frame with 4 variables}
\usage{
data(mxstates)
}
\description{
A fortified data.frame with Mexican state borders, to be passed to ggplot.
}
\details{
\itemize{
  \item long. Longitud
  \item lat. Latitud
  \item group. Group to identify same state coordinates
  \item id. Unique state id to match data.frames with information
}
}
\keyword{datasets}

